requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
dataclasses-json==0.6.7
Django==5.0.6
django-anymail==10.3
django-debug-toolbar==4.4.6
django-redis==5.4.0
djangorestframework==3.15.2
idna==3.7
jsonpath-python==1.0.6
kombu==5.3.7
marshmallow==3.21.3
mypy-extensions==1.0.0
ndg-httpsclient==0.5.1
packaging==24.1
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg-pool==3.2.2
pyasn1==0.6.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
shippo==3.6.1
six==1.16.0
sqlparse==0.5.0
stripe==9.12.0
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
